Move print_levels to new file, cleaner
